Enable indexOf tests
